Invoice Consolidation Script
I added an invoice consolidation script. The purpose here is that it will take the formularly record, read it, sum together similar components based on the item ID and return a consolidated, alphabetical list of medications, how many were issued, and the price per item. This version will be expanded to return items in the form of our medication records once our data structure for that is confirmed and built.